Main Objectives
Our primary focus points are getting those in recovery or  are still suffering from substance use disorder.
We fall under am organization 'CAPE TOWN DRUG COUBSELLING CENTRE M/PLAIN',, We using their building to do our work from.
what we doing currently. we running a feeding scheme, support groups, getting people into treatment for free through the change agents program. we have on a daily different motivational speaking to these members that need the encouragement and the speakers do it on a voluntary basis. we also have a sports day, we do it on a friday, and our primary focus sport currently is soccer because we only focusing on men for now.